backend/requirements.txt
python-jose[cryptography]
fastapi
uvicorn[standard]
httpx
python-dotenv
python-multipart